+++ R/cnd-abort.R
@@ -1333,7 +1333,7 @@ call_restore <- function(x, to) {
 #' * It prints the backtrace for the current error according to the
 #'   `rlang_backtrace_on_error` option.
 #'
-#' @section Errors in RMarkdown:
+#' @section Warnings and errors in RMarkdown:
 #'
 #' The display of errors depends on whether they're expected (i.e.
 #' chunk option `error = TRUE`) or unexpected:
@@ -1361,8 +1361,29 @@ call_restore <- function(x, to) {
 #' backtraces. If needed, you can override this by setting the
 #' `rlang_trace_top_env` global option.
 #'
+#' Similarly to `rlang_backtrace_on_error_report`, you can set
+#' `rlang_backtrace_on_warning_report` inside RMarkdown documents to
+#' tweak the display of warnings. This is useful in conjunction with
+#' [global_entrace()]. Because of technical limitations, there is
+#' currently no corresponding `rlang_backtrace_on_warning` option for
+#' normal R sessions.
+#'
+#' To get full entracing in an Rmd document, include this in a setup
+#' chunk before the first error or warning is signalled.
+#'
+#' ````
+#' ```{r setup}
+#' rlang::global_entrace()
+#' options(rlang_backtrace_on_warning_report = "full")
+#' options(rlang_backtrace_on_error_report = "full")
+#' ```
+#' ````
+#'
+#'
 #' @name rlang_backtrace_on_error
-#' @aliases add_backtrace
+#' @seealso rlang_backtrace_on_warning
+#' @aliases add_backtrace rlang_backtrace_on_error_report
+#'   rlang_backtrace_on_warning_report
 #'
 #' @examples
 #' # Display a simplified backtrace on error for both base and rlang
@@ -1452,6 +1473,22 @@ peek_backtrace_on_error <- function() {
 # by knitr
 peek_backtrace_on_error_report <- function() {
   peek_backtrace_on_error_opt("rlang_backtrace_on_error_report") %||% "none"
+}
+
+peek_backtrace_on_warning_report <- function() {
+  opt <- peek_backtrace_on_error_opt("rlang_backtrace_on_warning_report") %||% "none"
+
+  if (is_string(opt, "reminder")) {
+    options(rlang_backtrace_on_warning_report = "none")
+    warn(c(
+      "`rlang_backtrace_on_warning_report` must be one of `c(\"none\", \"branch\", \"full\")`.",
+      i = "The option was reset to \"none\"."
+    ))
+
+    opt <- "none"
+  }
+
+  opt
 }
 
 peek_backtrace_on_error_opt <- function(name) {