+++ R/cnd-entrace.R
@@ -22,6 +22,27 @@
 #' @param class A character vector of one or several classes of
 #'   conditions to be entraced.
 #'
+#' @section Inside RMarkdown documents:
+#'
+#' Call `global_entrace()` inside an RMarkdown document to cause
+#' errors and warnings to be promoted to rlang conditions that include
+#' a backtrace. This needs to be done in a separate setup chunk before
+#' the first error or warning.
+#'
+#' This is useful in conjunction with
+#' [`rlang_backtrace_on_error_report`] and
+#' [`rlang_backtrace_on_warning_report`]. To get full entracing in an
+#' Rmd document, include this in a setup chunk before the first error
+#' or warning is signalled.
+#'
+#' ````
+#' ```{r setup}
+#' rlang::global_entrace()
+#' options(rlang_backtrace_on_warning_report = "full")
+#' options(rlang_backtrace_on_error_report = "full")
+#' ```
+#' ````
+#'
 #' @section Under the hood:
 #' On R 4.0 and newer, `global_entrace()` installs a global handler
 #' with `globalCallingHandlers()`. On older R versions, `entrace()` is
@@ -36,7 +57,7 @@ global_entrace <- function(enable = TRUE,
   check_bool(enable)
   class <- arg_match(class, multiple = TRUE)
 
-  if (getRversion() < "4.0") {
+  if (getRversion() < "4.0" && !knitr_in_progress()) {
     return(global_entrace_fallback(enable, class))
   }
 
@@ -150,11 +171,25 @@ entrace <- function(cnd, ..., top = NULL, bottom = NULL) {
     return(entrace_handle_top(trace))
   }
 
-  # Log warnings and messages
+  # Log warnings
   if (is_warning(cnd)) {
-    push_warning(as_rlang_warning(cnd, trace))
-    return()
+    wrn <- as_rlang_warning(cnd, trace)
+    push_warning(wrn)
+
+    # Resignal enriched warning
+    if (!is_null(findRestart("muffleWarning"))) {
+      if (identical(peek_option("warn"), 2L)) {
+        return()
+      } else {
+        warning(wrn)
+        invokeRestart("muffleWarning")
+      }
+    } else {
+      return()
+    }
   }
+
+  # Log messages
   if (is_message(cnd)) {
     push_message(as_rlang_message(cnd, trace))
     return()