+++ tests/testthat/test-trace.R
@@ -463,7 +463,7 @@ test_that("caught error does not display backtrace in knitted files", {
   lines <- render_md("test-trace.Rmd")
   error_line <- lines[[length(lines)]]
   expect_match(error_line, "foo$")
-  
+
   expect_snapshot({
     cat_line(render_md("test-trace-full.Rmd"))
   })