+++ NEWS.md
@@ -1,5 +1,14 @@
 # rlang (development version)
 
+* `last_warnings()` and `last_messages()` now work inside RMarkdown
+  documents more accurately.
+
+* New `rlang_backtrace_on_warning_report` global option. This is
+  useful in conjunction with `global_entrace()` to get backtraces on
+  warnings inside RMarkdown documents.
+
+* `global_entrace()` now supports knitr documents.
+
 * `abort()`, `warn()`, and `inform()` gain an `.inherit` parameter.
   This controls whether `parent` is inherited. If `FALSE`,
   `cnd_inherits()` and `try_fetch()` do not match chained conditions